refactor(hooks): remove circular-dependency knot and gate cycles
Extract the hooks bridge routing table, parsed-config cache, and epoch
cells into a new leaf module, bridges/hooks/routing-state.ts, and
repoint the five knot importers at it. event-router.ts re-exports the
moved names so consumers outside the knot are unchanged.

Widen the local fallow gate to --circular-deps and --re-export-cycles
in the same commit, since the flags fail on the inherited cycles until
those cycles are gone. Closes FLOW-02.
