+++ .planning/codebase/ARCHITECTURE.md
@@ -99,7 +99,7 @@
 **bridges/:**
 - Purpose: one bridge per Claude-plugin component kind; each exposes `discover` (enumerate source artifacts + generated names), `stage`/`prepareStage*` (write into a staging dir or compute a prepared write), `commit*` (atomic rename/write into the live location), and `unstage*` (rollback removal)
 - Location: `extensions/pi-claude-marketplace/bridges/{skills,commands,agents,mcp,hooks}/`
-- Contains: kind-specific `types.ts`, `discover.ts`, `stage.ts`, `unstage.ts`, plus bridge-local helpers (e.g. `agents/frontmatter.ts`, `agents/index-mutation.ts`, `mcp/substitute.ts` for `${CLAUDE_PLUGIN_DATA}` variable substitution, `hooks/if-field/*` for the `if:` predicate compiler, `hooks/async-rewake/*` for hook-async resume state)
+- Contains: kind-specific `types.ts`, `discover.ts`, `stage.ts`, `unstage.ts`, plus bridge-local helpers (e.g. `agents/frontmatter.ts`, `agents/index-mutation.ts`, `mcp/substitute.ts` for `${CLAUDE_PLUGIN_DATA}` variable substitution, `hooks/if-field/*` for the `if:` predicate compiler, `hooks/async-rewake/*` for hook-async resume state, `hooks/routing-state.ts` — the leaf holding the hooks routing table, parsed-config cache, and epoch/pending-context cells, so the dispatch chain reads shared state without importing `event-router.ts` back)
 - Depends on: domain/ (name generation, manifest types), persistence/ (locations), shared/
 - Used by: orchestrators/plugin/* (5-phase install/uninstall ledgers)
 
@@ -208,7 +208,7 @@
 - **Circular imports:** Gated inside `orchestrators/` only, by two mechanisms that cover different halves of the problem (D-11):
   - `import-x/no-cycle` (BLOCK C-2 of `eslint.config.js`), scoped to `extensions/pi-claude-marketplace/orchestrators/**/*.ts`. The block also sets `import-x/extensions` to include `.ts`; without that setting the rule resolves imports but never parses the resolved `.ts` files, so it walks a one-node graph and greens on any cycle. `tests/architecture/import-boundaries.test.ts` pins both the rule and the setting.
   - A directed-edge grep gate in the same test file, because `no-cycle` fires only once a graph is *already* circular — the first of the two edges lands green. The ledger modules of `orchestrators/plugin/` (`install`/`update`/`uninstall`/`reinstall`/`enable-disable`) and of `orchestrators/marketplace/` (`add`/`remove`/`update`/`autoupdate`) must not import each other, in either direction, including `import type`. A plugin ledger reaches marketplace code only through `orchestrators/marketplace/shared.ts`; a marketplace file reaches plugin code only through leaf composers (`orchestrators/plugin/update-row.ts`, `clone-cache.ts`, `clone-gc.ts`), shared types in `orchestrators/types.ts`, or the injected `pluginUpdate` seam.
-  - Not covered: `orchestrators/plugin/bootstrap.ts` imports `marketplace/add.ts` and `marketplace/autoupdate.ts` by design — it is a composer, not a ledger. And `bridges/hooks/` carries a pre-existing cycle knot (`event-router.ts` ↔ `dispatch.ts` ↔ `async-rewake/registry.ts`), which is why the `no-cycle` glob stops at `orchestrators/` rather than covering the whole extension.
+  - Not covered: `orchestrators/plugin/bootstrap.ts` imports `marketplace/add.ts` and `marketplace/autoupdate.ts` by design — it is a composer, not a ledger. The `bridges/hooks/` cycle knot (`event-router.ts` ↔ `dispatch.ts` ↔ `async-rewake/registry.ts`) that used to sit outside the glob was removed by extracting the bridge's shared module state — routing table, parsed-config cache, epoch and pending-context cells — into the `bridges/hooks/routing-state.ts` leaf; whole-repo cycle coverage now comes from `npm run fallow`'s `--circular-deps` flag rather than from the ESLint glob. Whether `import-x/no-cycle` should now widen past `orchestrators/` is an open question (BACKLOG FLOW-03), deliberately out of scope of the knot removal because widening the glob has its own blast radius across the extension.
 - **Network boundary:** `orchestrators/plugin/install.ts`, `list.ts`, `uninstall.ts` MUST NOT import `platform/git.ts` or carry a `gitOps` field — enforced by `tests/architecture/no-orchestrator-network.test.ts`, a source-grep architectural test (NFR-5)
 - **Lock re-entrancy:** `proper-lockfile` is configured `retries: 0` and is NOT re-entrant; nesting two `withLockedStateTransaction` calls on the same scope's lock file self-deadlocks (`ELOCKED` → `StateLockHeldError`) — guard-free ledger bodies (`runInstallLedger`, etc.) exist specifically so callers that already hold the lock (e.g. `setPluginEnabled`'s enable branch) can invoke the ledger without re-acquiring
